requirements.txt
requests
bs4
discord.py
python-dotenv
jsonschema